requirements.txt
sounddevice
numpy
scipy
groq
pynput
openai
deepgram-sdk
python-dotenv